copy oracle template
