+++ package.json
@@ -34,6 +34,7 @@
     "jsonapi-serializer": "^3.5.6",
     "lodash": "^4.17.11",
     "moment": "^2.22.2",
+    "oracledb": "^3.0.0",
     "simple-git": "^1.96.0",
     "winston": "^3.1.0",
     "winston-daily-rotate-file": "^3.3.2"
